[Fix] : 내 게시글 받아오는 함수 Impl로 이동
