+++ app/src/main/java/com/link_up/matching_manager/domain/repository/MyRepository.kt
@@ -1,23 +1,24 @@
 package com.link_up.matching_manager.domain.repository
 
+import androidx.lifecycle.MutableLiveData
 import com.link_up.matching_manager.ui.match.MatchDataModel
 import com.link_up.matching_manager.ui.team.TeamItem
 import com.google.firebase.database.FirebaseDatabase
 
 interface MyRepository {
-    suspend fun getMatchList(userId : String, database: FirebaseDatabase): List<MatchDataModel>
+    fun getMatchList(database: FirebaseDatabase, list : MutableLiveData<List<MatchDataModel>>)
 
     suspend fun deleteMatchData(data : MatchDataModel, database: FirebaseDatabase)
 
     suspend fun editMatchData(data : MatchDataModel, newData: MatchDataModel, database: FirebaseDatabase)
 
-    suspend fun getRecruitList(userId : String, database: FirebaseDatabase): List<TeamItem.RecruitmentItem>
+    fun getRecruitList(database: FirebaseDatabase, list : MutableLiveData<List<TeamItem.RecruitmentItem>>)
 
     suspend fun deleteRecruitData(data : TeamItem.RecruitmentItem, database: FirebaseDatabase)
 
     suspend fun editRecruitData(data : TeamItem.RecruitmentItem, newData: TeamItem.RecruitmentItem, database: FirebaseDatabase)
 
-    suspend fun getApplicationList(userId : String, database: FirebaseDatabase): List<TeamItem.ApplicationItem>
+    fun getApplicationList(database: FirebaseDatabase, list : MutableLiveData<List<TeamItem.ApplicationItem>>)
 
     suspend fun deleteApplicationData(data : TeamItem.ApplicationItem, database: FirebaseDatabase)
